% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbox.R
\name{test_function}
\alias{test_function}
\title{Test function}
\usage{
test_function(a = 1, b = 2)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number.
}
\description{
This is a test function.
}
