% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbox.R
\name{test_function2}
\alias{test_function2}
\title{Test function 2}
\usage{
test_function2(a = 1, b = 2, c = 3)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}

\item{c}{A number.}
}
\value{
A number.
}
\description{
This is a test function 2.
}
